Create stage setup for auth
