+++ client/src/components/index.ts
@@ -1,3 +1,4 @@
 // eslint-disable-next-line
 export * from './shared';
 export * from './templates';
+export * from './signup';
