+++ client/src/components/signup/index.tsx
@@ -1,0 +1,2 @@
+// eslint-disable-next-line
+export {default as Signup} from './Signup';
